preserve temp dir while db in use
for tests, we create a temp dir to temporarily hold the db
contents, before this fix the temp directory was created but
dropped before we were done with the database.

When inserting a lot of documents additional log files will get
created, becuase the tmp dir was dropped creation failed.

This fix preserves the temp dir until we are done with operations
preventing creation error.